requirements.txt
boto==2.38.0
defusedxml==0.4.1
Django==1.8.2
django-admin-bootstrapped==2.5.4
django-allauth==0.20.0
django-bootstrap3==6.1.0
django-cors-headers==1.1.0
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-filter==0.10.0
django-guardian==1.3
django-model-utils==2.2
django-rest-auth==0.4.0
django-storages-redux==1.2.3
django-summernote==0.6.8
djangorestframework==3.1.3
Markdown==2.6.2
oauthlib==0.7.2
Pillow==2.8.2
psycopg2==2.6.1
python3-openid==3.0.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.15
